stocks/serializers.py
from rest_framework import serializers
from .models import Supplier


class SupplierSerializer(serializers.ModelSerializer):
    class Meta:
        model = Supplier
        fields = ('user', 'name', 'address' ,'created_date')
